Fix string vs byte errors in test_object_services
The object service tests was using strings for the data payload in
tests of the object service. In python 3, the client returns these
objects as bytes instead of strings, causing comparisons to fail. This
changes the tests to bytes for the data types instead of strings, so
that the comparisons to the returned bytes will work in both python 2
and python 3.

Change-Id: I896a5fc6cef1d19439327f587c59670473632927